pyproject.toml
[tool.poetry]
name = "fastapi_boilerplate"
version = "0.1.0"
description = "fastapi boilerplate"
authors = ["valerybriz <valerybriz@gmail.com"]
license = "MIT"


[[tool.poetry.source]]
name = "public pypi"
url = "https://pypi.python.org/simple"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.6"
gunicorn = "^20.0.4"
fastapi = "^0.59.0"
passlib = "^1.7.2"
pyjwt = "^1.7.1"
python-slugify = "^4.0.1"
Unidecode = "^1.1.1"
python-dotenv = "^0.14.0"
databases = "^0.3.2"
motor = "^2.1.0"
bcrypt = "^3.1.7"
pydantic = {extras = ["email"], version = "^1.6.1"}
sentry-sdk = "^0.17.3"
pyyaml = "^5.4.1"
requests = "^2.25.1"


[tool.poetry.dev-dependencies]
pytest = "*"
pre-commit = "^2.6.0"
flake8 = "^3.8.3"
coverage = "^5.2.1"
black = "^19.10b0"
bandit = "^1.6.2"

[tool.black]
line-length = 80
max-line-length = 80
exclued = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
